+++ src/main/java/com/blame/artneunet/network/activation/IdentityActivationFunction.java
@@ -3,7 +3,7 @@ package com.blame.artneunet.network.activation;
 public class IdentityActivationFunction implements IActivationFunction {
 
 	@Override
-	public float calculateActivationValue(float processedValue, int numOfInConnections) {
+	public double calculateActivationValue(double processedValue, int numOfInConnections) {
 		return processedValue;
 	}
