+++ src/main/java/com/blame/artneunet/network/neuron/ProcessingNeuron.java
@@ -47,7 +47,7 @@ public class ProcessingNeuron extends Neuron {
 	 */
 	public void process() {
 		
-		float processingValue = 0f;
+		double processingValue = 0d;
 		for(Connection inConnection : inConnections) {
 			processingValue += inConnection.getInputNeuron().getValue() * inConnection.getWeight();
 		}
