+++ src/test/java/com/blame/artneunet/network/NetworkTest.java
@@ -5,8 +5,6 @@ import java.util.List;
 
 import org.junit.Test;
 
-import com.blame.artneunet.network.Network;
-
 public class NetworkTest {
 
 	@Test
@@ -14,9 +12,9 @@ public class NetworkTest {
 		
 		Network network = new Network(2, Arrays.asList(4), 2);
 		network.randomizeConnections();
-		network.setInputLayerValues(Arrays.asList(0.2f, 0.8f));
+		network.setInputLayerValues(Arrays.asList(0.2d, 0.8d));
 		network.processNetwork();
-		List<Float> outputLayerValues = network.getOutputLayerValues();
+		List<Double> outputLayerValues = network.getOutputLayerValues();
 		System.out.println(outputLayerValues);
 	}
 }
